wasdfklj

+++ proj2/src/vm/frame.h
@@ -8,10 +8,10 @@
 struct list frame_table;
 
 struct frame_table_entry {
-	uint32_t* frame;
+	uint8_t* frame;
 	struct thread* owner;
 	struct sup_page_table_entry* aux;
 	struct list_elem elem;
 };
 
-struct frame_table_entry* frame_table_insert(uint32_t* frame, struct thread* owner, struct sup_page_table_entry* aux);
+struct frame_table_entry* frame_table_insert(uint8_t* frame, struct thread* owner, struct sup_page_table_entry* aux);
