+++ proj2/src/vm/page.h
@@ -6,7 +6,7 @@
 
 
 struct sup_page_table_entry {
-	uint32_t* uservaddr;
+	uint8_t* uservaddr;
 	uint64_t access_time;
 	int block_index;
 	bool dirty;
@@ -14,4 +14,4 @@ struct sup_page_table_entry {
 	struct list_elem elem;
 };
 
-struct sup_page_table_entry* sup_page_table_insert(uint32_t* uservaddr, uint64_t access_time, int block_index, bool dirty, bool accessed);
+struct sup_page_table_entry* sup_page_table_insert(uint8_t* uservaddr, uint64_t access_time, int block_index, bool dirty, bool accessed);
